+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # v5.0.0
 
 - FEAT: Auto & manual device selection via `notion.selectDevice(...)` method
-- FEAT: new methods: `notion.getDevices()` and `notion.getSelectedDevice()`
+- FEAT: new method: `notion.getDevices()`
 - FIX: #46 Notion sends 1 packet of data even though it is asleep
 - FIX: only send timesync actions if and when device is online